feat: multiplatform mutators
