+++ generators/build.gradle.kts
@@ -24,8 +24,6 @@ sourdough {
 }
 
 dependencies {
-  api(projects.satisfaketionCore)
-  implementation(group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-json", version = "1.3.2")
   detektPlugins(group = "io.gitlab.arturbosch.detekt", name = "detekt-formatting", version = "1.19.0")
 }
